requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
docopt==0.6.2
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
ptyprocess==0.7.0
Pygments==2.19.2
pytest==8.4.1
pytest-watch==4.2.0
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
watchdog==6.0.0
